Fix library structure
Previously, the library had an invalid format, where the library.properties metadata file was in the correct location in
the root of the library, but the header file and examples were in an `animationTools` subfolder.

Moving everything up out of the `animationTools` folder to the root of the repository results in a valid library structure:

https://arduino.github.io/arduino-cli/latest/library-specification/#layout-of-folders-and-files

This will allow the library to be accepted into the Arduino Library Manager.