Bump version in library metadata to 0.6
In order to allow the library to be added to Library Manager, a restructuring of the repository was done. A new release of the library must now be made to distribute a valid version of the library.

In preparation for the release, the `version` value in the `library.properties` metadata file must be updated.